old/5-langchain-rag-intro/requirements.txt
pypdf==6.7.5
chromadb==1.5.1
langchain-chroma==1.1.0
langchain_google_genai==4.2.1
langchain_community==0.4.1
langchain_text_splitters==1.1.1